Fix iPhone X / XS Bug
On the iPhone X / XS, the cropView and fadeView were not perfectly on top of each other (offset in y direction) and in addition, the Crop / Retake buttons weren't pinned to the bottom of the screen. This is fixed by the first couple lines.  
I also set the height / width constants not to be proportional to the screen size but fixed since that's what I generally prefer for this situation.

You'll need the following extension for the first few lines to work:

extension UIView {
    func addConstraintsWithFormat(format: String, views: UIView...) {
        var viewDict = [String: AnyObject]()
        for (index, view) in views.enumerated() {
            view.translatesAutoresizingMaskIntoConstraints = false
            let key = "v\(index)"
            viewDict[key] = view
        }
        addConstraints(NSLayoutConstraint.constraints(withVisualFormat: format, options: NSLayoutConstraint.FormatOptions(), metrics: nil, views: viewDict))
    }
}